+++ lib/proto_base_client.dart
@@ -1,3 +1,4 @@
 library proto_base_client;
 
 export 'src/api_client.dart';
+export 'src/database_client.dart';