+++ CHANGELOG.md
@@ -2,3 +2,10 @@
 - Added API tokens to validate application requests.
 - Users can enroll for API keys from the [official website](https://protobase.pythonanywhere.com/).
 - Updated Documentation
+
+## 1.0.2 - Version 2.0.0
+- Added `ProtoBaseDatabaseClient` for database operations.
+  - Methods for creating tables, inserting data, reading data, updating data, and deleting data.
+- Improved error handling in API client methods.
+- Added example usage in `main.dart`.
+- Updated documentation with new API methods.